Refactor hotkey
