+++ lib/settings/settings.dart
@@ -1,4 +1,5 @@
 export 'cubit/cubit.dart';
+export 'hotkey.dart';
 export 'icon_manager.dart';
 export 'settings_page.dart';
 export 'settings_service.dart';
